Fix stop sound omx

+++ cueList.js
@@ -2,12 +2,12 @@
 Copyright ou © ou Copr. Clément Bossut, (2018)
 <bossut.clement@gmail.com>
 
-Ce logiciel est un programme informatique servant à écrire et jouer une conduite lumière synchronisée avec du son sur une Raspberry Pi avec PCA8596. 
+Ce logiciel est un programme informatique servant à écrire et jouer une conduite lumière synchronisée avec du son sur une Raspberry Pi avec PCA8596.
 
 Ce logiciel est régi par la licence CeCILL soumise au droit français et
 respectant les principes de diffusion des logiciels libres. Vous pouvez
 utiliser, modifier et/ou redistribuer ce programme sous les conditions
-de la licence CeCILL telle que diffusée par le CEA, le CNRS et l'INRIA 
+de la licence CeCILL telle que diffusée par le CEA, le CNRS et l'INRIA
 sur le site "http://www.cecill.info".
 */
 
@@ -30,6 +30,7 @@ let cl = {
   , player = null
   , soundTimeRef = -1
   , soundPosRef = 0
+  , playerALSA = false
 
 cl.new = function() {
   this.soundPath = ''
@@ -88,7 +89,7 @@ cl.play = function(pos = 0, cb = null) {
   let second = Math.floor(pos/1000)
   soundPosRef = second*1000
   playTimeouts = []
-  
+
   let lastDateIndex = 0
   for(let i = 0 ; i < this.content.length ; i++) {
     if (this.content[i].date > pos) break;
@@ -96,9 +97,15 @@ cl.play = function(pos = 0, cb = null) {
   }
   this.applyCue(lastDateIndex)
   let offset = soundPosRef - this.content[lastDateIndex].date
-  
-  if (second) player = spawn('omxplayer', [this.soundPath, '-Il', second.toString()])
-  else player = spawn('aplay', ['-Dhw:CARD=XDA2', this.soundPath, '-v'])
+
+  if (second) {
+    player = spawn('omxplayer', [this.soundPath, '-Il', second.toString()])
+    playerALSA = false
+  }
+  else {
+    player = spawn('aplay', ['-Dhw:CARD=XDA2', this.soundPath, '-v'])
+    playerALSA = true
+  }
   player.stderr.once('data', () => {
     soundTimeRef = new Date().getTime()
     this.content.forEach((v,i,a)=>{
@@ -112,9 +119,13 @@ cl.play = function(pos = 0, cb = null) {
 cl.cut = function() {
   if (!player) return;
   this.stop()
-  player.stdin.write(' ')
+  if (playerALSA) {
+    player.stdin.write(' ')
+  }
   player.stdin.write('q')
-  player.kill('SIGINT')
+  if (playerALSA) {
+    player.kill('SIGINT')
+  }
   soundPosRef = new Date().getTime() - soundTimeRef + soundPosRef
   soundTimeRef = -1
 }
@@ -148,7 +159,7 @@ cl.go = function(n = 0, cb = null, offset = 0) {
   }
   let downMs = downDiff == -1 ? 0 : this.content[n].downTime * 1000
     , upMs = upDiff == -1 ? 0 : this.content[n].upTime * 1000
-  
+
   this.interMs = Math.max(
     this.orgue.minMsInter,
     Math.min(downMs/downDiff, upMs/upDiff)
@@ -156,11 +167,11 @@ cl.go = function(n = 0, cb = null, offset = 0) {
   this.downFrames = Math.round(downMs/this.interMs)
   this.upFrames = Math.round(upMs/this.interMs)
   this.frameCount = 0
-  
+
   let start = this.from.slice()
   if (!this.downFrames) this.downLeds.forEach(v => start[v] = this.to[v])
   if (!this.upFrames) this.upLeds.forEach(v => start[v] = this.to[v])
-  
+
   this.orgue.state = start
   this.nextFrame = new Date().getTime() - offset
   this.timeOut = setTimeout(()=>{this.inter(cb)}, this.interMs)
@@ -177,7 +188,7 @@ cl.inter = function(cb) {
     this.nextFrame += jumped * this.interMs
     this.frameCount += jumped
   }
-  
+
   let loop = false
     , ratioUp = this.frameCount/this.upFrames
     , ratioDown = this.frameCount/this.downFrames
@@ -199,7 +210,7 @@ cl.inter = function(cb) {
     this.downLeds.forEach(v => res[v] = this.to[v])
   }
   this.orgue.state = res
-  
+
   if (loop) this.timeOut = setTimeout(()=>{this.inter(cb)}, this.interMs - diff)
   if (cb) cb(loop, this.frameCount*this.interMs)
 }
